requirements.txt
langchain==1.1.3
langchain-core==1.2.13
langchain-openai==1.1.3
langchain-xai==1.1.0
langchain-anthropic==1.3.3
pandas==2.3.3
python-dotenv==1.2.1
streamlit==1.52.1
pydantic==2.12.5
openai==2.11.0
pymupdf==1.26.7
nest_asyncio==1.6.0
pymongo==4.16.0
anthropic==0.79.0